Тесты, исправления UserParametersWorker'а

+++ app/src/main/java/korablique/recipecalculator/database/FoodstuffsDbHelper.java
@@ -93,7 +93,10 @@ public class FoodstuffsDbHelper {
 
     public static synchronized void deinitializeDatabase(Context context) {
         File dbFile = getDbFile(context);
-        boolean deleted = dbFile.delete();
+        boolean deleted = true;
+        if (dbFile.exists()) {
+            deleted = dbFile.delete();
+        }
         if (!deleted) {
             throw new Error("Couldn't delete database");
         }
